Improve 'Delete Category' option

+++ js/tasksFunctions.js
@@ -1,4 +1,8 @@
 const functions = {
+    /**
+     * Add a new category to the list of categories displayed on the webpage. 
+     * @param {*} categoryName 
+     */
     addCategory : (categoryName) => {
         const categories = document.querySelector("#categories");
         const newCategory = document.createElement("article");
@@ -8,18 +12,17 @@ const functions = {
             <form action="#" method="POST" onsubmit="return false">
                 <input type="text" name="newTask" value="" placeholder="New task">
                 <input type="submit" class="addTask" value="Add Task">                
-                <input type="submit" class="deleteCategory" value="Delete Category">
+                <input id="${categoryName.trim().replaceAll(" ", "")}" type="submit" class="deleteCategory" value="Delete Category">
             </form>`;
-        
-        newCategory.id = categoryName.trim().replaceAll(" ", "");
+                
         categories.appendChild(newCategory);
 
         // Trim category field
         document.querySelector("#categoryName").value = "";
 
-        // Add event listener to delete category button
+        // Add event listener to delete category button of the new category
         newCategory.querySelector(".deleteCategory").addEventListener("click", () => {
-            functions.deleteCategory();
+            functions.deleteCategory(categoryName.trim().replaceAll(" ", ""));
         });
 
         // Save category to localStorage
@@ -28,24 +31,29 @@ const functions = {
         localStorage.setItem("categories", JSON.stringify(categoriesList));
     },
 
-    deleteCategory : () => {
-        //! Improve this function
-        const categoryElement = event.target.closest("article");
-        const categoryName = categoryElement.querySelector("h3").textContent;
-        const categoriesList = JSON.parse(localStorage.getItem("categories")) || [];
-        categoriesList.splice(categoriesList.indexOf(categoryName), 1);
-        localStorage.setItem("categories", JSON.stringify(categoriesList));
-        categoryElement.remove();
+    /**
+     * Delete a category from the list of categories displayed on the webpage
+     * @param {*} id 
+     */
+    deleteCategory : (id) => {            
+        if(id) {            
+            // If the user doesn't want to delete the category, return
+            let response = confirm("Are you sure you want to delete this category?");
+            if(!response) return;
 
-        
-        /* const categories = document.querySelector("#categories");
-        const categoryName = categories.querySelector("h3").textContent;        
-        const categoriesList = JSON.parse(localStorage.getItem("categories")) || [];
-        categoriesList.splice(categoriesList.indexOf(categoryName), 1);
-        localStorage.setItem("categories", JSON.stringify(categoriesList));                
-        categories.removeChild(categories.querySelector("#" + categoryName)); */
+            const categoryElement = document.querySelector("#" + id).closest("article");            
+            const categoriesList = JSON.parse(localStorage.getItem("categories")) || [];
+
+            categoriesList.splice(categoriesList.indexOf(id), 1);
+            localStorage.setItem("categories", JSON.stringify(categoriesList));
+            
+            categoryElement.remove();
+        }             
     },
 
+    /**
+     * Show all categories on the webpage
+     */
     showCategories : () => {
         const categories = document.querySelector("#categories");
         const categoriesList = JSON.parse(localStorage.getItem("categories")) || [];
@@ -56,7 +64,7 @@ const functions = {
                 <form action="#" method="POST" onsubmit="return false">
                     <input type="text" name="newTask" value="" placeholder="New task">
                     <input type="submit" class="addTask" value="Add Task">                
-                    <input type="submit" class="deleteCategory" value="Delete Category">
+                    <input id="${category.trim().replaceAll(" ", "")}" type="submit" class="deleteCategory" value="Delete Category">
                 </form>`;
             categories.appendChild(newCategory);
         });
